==> 2. modeling/28_add_velocity_linear_interpolation_vector_filed_polar_coordinate.ipynb <==
import pandas as pd
import numpy as np
import math
from scipy.spatial import cKDTree
from scipy.optimize import curve_fit
import warnings

warnings.filterwarnings("ignore")

main = pd.read_csv("data/total.csv")
water = pd.read_csv("data/water_velocity.csv")
wind = pd.read_csv("data/wind_velocity.csv")

water = water.fillna(0)
wind = wind.fillna(0)

main["water_u"] = 0
main["water_v"] = 0
main["wind_u"] = 0
main["wind_v"] = 0
main['d_long'] = main['longitude'].diff().fillna(0)
main['d_lati'] = main['latitude'].diff().fillna(0)

main = main.groupby('drifter').apply(lambda group: group.iloc[1:]).reset_index(drop=True)

# Create a grid of longitude and latitude from the water DataFrame
grid = water[['longitude', 'latitude']].drop_duplicates().values

# Create a KD-Tree for efficient nearest neighbor search
kdtree = cKDTree(grid)

# Function to find the 4 closest vertices for a given point
def find_closest_vertices(row):
    point = np.array([row['longitude'], row['latitude']])
    # Query the KD-Tree to find the 4 closest vertices
    _, indices = kdtree.query(point, k=4)
    return indices

# Apply the function to the main DataFrame to find the 4 closest vertices for each point
main['closest_vertices'] = main.apply(find_closest_vertices, axis=1)

main

water

r = []
theta = []
for i in range(len(main)):
    x = water.loc[water["time"] == main.iloc[i][1]]
    w_d = []
    w_r = []
    w_theta = []
    
    for j in main.iloc[i][-1]:
        d = ((main.iloc[i][2] - x.iloc[j][1])**2 + (main.iloc[i][3] - x.iloc[j][2])**2) ** 0.5
        _r = ((x.iloc[j][-2] ** 2) + (x.iloc[j][-1] ** 2)) * 0.5
        
        u = np.array([x.iloc[j][-2], 0])
        v = np.array([0, x.iloc[j][-1]])
        dot_product = np.dot(u, u + v)
        u_norm = np.linalg.norm(u)
        u_plus_v_norm = np.linalg.norm(u+v)
        
        cos_theta = dot_product / (u_norm * u_plus_v_norm)
        _theta = math.acos(cos_theta)
        
        w_d.append(d)
        w_r.append(_r)
        w_theta.append(_theta)
        
    weights = [1 / d for d in w_d]
    total_weight = sum(weights)
    normalized_weights = [weight/total_weight for weight in weights]
    
    interpolated_r = sum(weight * r for weight, r in zip(normalized_weights, w_r))
    interpolated_theta = sum(weight * theta for weight, theta in zip(normalized_weights, w_theta))
    
    r.append(interpolated_r)
    theta.append(interpolated_theta)

w_u = []
w_v = []

for i in range(len(r)):
    u = r[i] * math.cos(theta[i])
    v = r[i] * math.sin(theta[i])
    w_u.append(u)
    w_v.append(v)
    
main["water_u"] = w_u
main["water_v"] = w_v

main

main.describe()

# Create a grid of longitude and latitude from the water DataFrame
grid2 = wind[['longitude', 'latitude']].drop_duplicates().values

# Create a KD-Tree for efficient nearest neighbor search
kdtree = cKDTree(grid2)

# Function to find the 4 closest vertices for a given point
def find_closest_vertices(row):
    point = np.array([row['longitude'], row['latitude']])
    # Query the KD-Tree to find the 4 closest vertices
    _, indices = kdtree.query(point, k=4)
    return indices

# Apply the function to the main DataFrame to find the 4 closest vertices for each point
main['closest_vertices2'] = main.apply(find_closest_vertices, axis=1)

r = []
theta = []
for i in range(len(main)):
    x = wind.loc[wind["time"] == main.iloc[i][1]]
    w_d = []
    w_r = []
    w_theta = []
    
    for j in main.iloc[i][-1]:
        d = ((main.iloc[i][2] - x.iloc[j][1])**2 + (main.iloc[i][3] - x.iloc[j][2])**2) ** 0.5
        _r = ((x.iloc[j][-2] ** 2) + (x.iloc[j][-1] ** 2)) * 0.5
        
        u = np.array([x.iloc[j][-2], 0])
        v = np.array([0, x.iloc[j][-1]])
        dot_product = np.dot(u, u + v)
        u_norm = np.linalg.norm(u)
        u_plus_v_norm = np.linalg.norm(u+v)
        
        cos_theta = dot_product / (u_norm * u_plus_v_norm)
        _theta = math.acos(cos_theta)
        
        w_d.append(d)
        w_r.append(_r)
        w_theta.append(_theta)
        
    weights = [1 / d for d in w_d]
    total_weight = sum(weights)
    normalized_weights = [weight/total_weight for weight in weights]
    
    interpolated_r = sum(weight * r for weight, r in zip(normalized_weights, w_r))
    interpolated_theta = sum(weight * theta for weight, theta in zip(normalized_weights, w_theta))
    
    r.append(interpolated_r)
    theta.append(interpolated_theta)

for i in range(10):
    print(r[i])

w_u = []
w_v = []

for i in range(len(r)):
    u = r[i] * math.cos(theta[i])
    v = r[i] * math.sin(theta[i])
    w_u.append(u)
    w_v.append(v)
    
main["wind_u"] = w_u
main["wind_v"] = w_v

main

main.to_csv("data/total_add_velocity_linear_interpolation_vector_field_polar_coordinate.csv", index = False)

main.describe()
