==> 2. modeling/30_compare_data(2).ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

weigthed_average_polar_coordinate = pd.read_csv("data/total_add_velocity_weigthed_average_polar_coordinate.csv")
bilinear_interpolation_polar_coordinate = pd.read_csv("data/total_add_velocity_bilinear_interpolation_polar_coordinate.csv")

df = weigthed_average_polar_coordinate

# 'u'와 'v' 열의 분포 시각화
plt.figure(figsize=(12, 6))  # 그래프의 크기 설정

# 'u' 열의 히스토그램
plt.subplot(1, 2, 1)
sns.histplot(df['water_u'], bins=30, kde=True)
plt.title('Weighted averge water u')
plt.xlabel('u')

# 'v' 열의 히스토그램
plt.subplot(1, 2, 2)
sns.histplot(df['water_v'], bins=30, kde=True)
plt.title('Weighted averge water v')
plt.xlabel('v')

plt.tight_layout()
plt.show()

df = weigthed_average_polar_coordinate

# 'u'와 'v' 열의 분포 시각화
plt.figure(figsize=(12, 6))  # 그래프의 크기 설정

# 'u' 열의 히스토그램
plt.subplot(1, 2, 1)
sns.histplot(df['wind_u'], bins=30, kde=True)
plt.title('Weighted averge wind u')
plt.xlabel('u')

# 'v' 열의 히스토그램
plt.subplot(1, 2, 2)
sns.histplot(df['wind_v'], bins=30, kde=True)
plt.title('Weighted averge wind v')
plt.xlabel('v')

plt.tight_layout()
plt.show()

df = weigthed_average_polar_coordinate

# Visualize the distribution of 'u' and 'v' columns with boxplots
plt.figure(figsize=(12, 12))  # Set the size of the graph

# Boxplot of column 'u'
plt.subplot(2, 2, 1)
sns.boxplot(data=df, y='water_u')
plt.title('Weighted average water u')
plt.ylabel('water u')

# Boxplot of column 'v'
plt.subplot(2, 2, 2)
sns.boxplot(data=df, y='water_v')
plt.title('Weighted average water v')
plt.ylabel('water v')

# Boxplot of column 'u'
plt.subplot(2, 2, 3)
sns.boxplot(data=df, y='wind_u')
plt.title('Weighted average wind u')
plt.ylabel('wind u')

# Boxplot of column 'v'
plt.subplot(2, 2, 4)
sns.boxplot(data=df, y='wind_v')
plt.title('Weighted average wind v')
plt.ylabel('wind v')

plt.tight_layout()
plt.show()

df = bilinear_interpolation_polar_coordinate

# 'u'와 'v' 열의 분포 시각화
plt.figure(figsize=(12, 6))  # 그래프의 크기 설정

# 'u' 열의 히스토그램
plt.subplot(1, 2, 1)
sns.histplot(df['water_u'], bins=30, kde=True)
plt.title('Weighted averge water u')
plt.xlabel('u')

# 'v' 열의 히스토그램
plt.subplot(1, 2, 2)
sns.histplot(df['water_v'], bins=30, kde=True)
plt.title('Weighted averge water v')
plt.xlabel('v')

plt.tight_layout()
plt.show()

df = bilinear_interpolation_polar_coordinate

# 'u'와 'v' 열의 분포 시각화
plt.figure(figsize=(12, 6))  # 그래프의 크기 설정

# 'u' 열의 히스토그램
plt.subplot(1, 2, 1)
sns.histplot(df['wind_u'], bins=30, kde=True)
plt.title('Weighted averge wind u')
plt.xlabel('u')

# 'v' 열의 히스토그램
plt.subplot(1, 2, 2)
sns.histplot(df['wind_v'], bins=30, kde=True)
plt.title('Weighted averge wind v')
plt.xlabel('v')

plt.tight_layout()
plt.show()

df = bilinear_interpolation_polar_coordinate

# Visualize the distribution of 'u' and 'v' columns with boxplots
plt.figure(figsize=(12, 12))  # Set the size of the graph

# Boxplot of column 'u'
plt.subplot(2, 2, 1)
sns.boxplot(data=df, y='water_u')
plt.title('Weighted average water u')
plt.ylabel('water u')

# Boxplot of column 'v'
plt.subplot(2, 2, 2)
sns.boxplot(data=df, y='water_v')
plt.title('Weighted average water v')
plt.ylabel('water v')

# Boxplot of column 'u'
plt.subplot(2, 2, 3)
sns.boxplot(data=df, y='wind_u')
plt.title('Weighted average wind u')
plt.ylabel('wind u')

# Boxplot of column 'v'
plt.subplot(2, 2, 4)
sns.boxplot(data=df, y='wind_v')
plt.title('Weighted average wind v')
plt.ylabel('wind v')

plt.tight_layout()
plt.show()
