==> 2. modeling/31_water_velocity2.ipynb <==
import pandas as pd
import netCDF4 as nc

# NetCDF 파일 열기
file_path = "data/water_velocity2016.nc"
dataset = nc.Dataset(file_path, "r")

# 파일 내의 변수 이름 확인
variables = dataset.variables.keys()
print(variables)

time = pd.DataFrame(dataset["time"][:])
for i in range(len(time)):
    time.iloc[i] = 0.125 * i
time

2016년 윤년이라 2월달에 29일까지 있음...

vo = pd.DataFrame(dataset["vo"][0][0][:])
vo

uo = pd.DataFrame(dataset["uo"][0][0][:])
uo

lat = pd.DataFrame(dataset["latitude"][:])
lat

lon = pd.DataFrame(dataset["longitude"][:])
lon

t = [i for i in range(int(0 * 100), int((730.75 + 0.25) * 100), int(0.25 * 100)) for _ in range(8740)]
t = [x / 100 for x in t]

lati = []
for i in range(len(lat)):
    cnt = [lat.iloc[i][0]] * 92
    lati += cnt
lati *= 2924

long = []
for i in range(len(lon)):
    long.append(lon.iloc[i][0])
long *= 95
long *= 2924

u_lst = []
for i in range(2924):
    x = pd.DataFrame(dataset["uo"][i * 2][0][:])
    new_x = x.values.flatten().tolist()
    u_lst.append(new_x)

v_lst = []
for i in range(2924):
    x = pd.DataFrame(dataset["vo"][i * 2][0][:])
    new_x = x.values.flatten().tolist()
    v_lst.append(new_x)
    
def flatten(lst):
    flattened_lst = []
    for sublist in lst:
        for item in sublist:
            flattened_lst.append(item)
    return flattened_lst

u_lst = flatten(u_lst)
v_lst = flatten(v_lst)

print(len(t))
print(len(lati))
print(len(long))
print(len(u_lst))
print(len(v_lst))

2924 * 95 * 92

data = {
    "time" : t,
    "longitude" : long,
    "latitude" : lati,
    "u" : u_lst,
    "v" : v_lst
}
columns = ["time","longitude","latitude","u","v"]
total = pd.DataFrame(data, columns = columns)
total

total.describe()

total.info()

total.to_csv("data/water_velocity2016.csv", index = False)
