==> 2. modeling/34_fit_100days.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import math
from datetime import datetime

from scipy.spatial import cKDTree
from scipy.optimize import curve_fit
import lightgbm as lgb
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error

import warnings

warnings.filterwarnings("ignore")

data = pd.read_csv("data/total_add_velocity_weighted_average.csv")
target = pd.read_csv("data/target.csv")
water = pd.read_csv("data/water_velocity2016.csv")
wind = pd.read_csv("data/wind_velocity2016.csv")

# Make predict model

reg = data.copy()
reg = reg.iloc[:,4:10]

reg

features = ['water_u', 'water_v', 'wind_u', 'wind_v']
target_d_long = 'd_long'
target_d_lati = 'd_lati'

X_train, X_test, y_train_long, y_test_long, y_train_lati, y_test_lati = train_test_split(
    reg[features], reg[target_d_long], reg[target_d_lati], test_size=0.2, random_state=42)

model_d_long = lgb.LGBMRegressor()
model_d_lati = lgb.LGBMRegressor()

model_d_long.fit(X_train, y_train_long)
model_d_lati.fit(X_train, y_train_lati)

predictions_d_long = model_d_long.predict(X_test)
predictions_d_lati = model_d_lati.predict(X_test)

rmse_d_long = math.sqrt(mean_squared_error(y_test_long, predictions_d_long))
rmse_d_lati = math.sqrt(mean_squared_error(y_test_lati, predictions_d_lati))

print(f'RMSE for d_long: {rmse_d_long}')
print(f'RMSE for d_lati: {rmse_d_lati}')

# Preprocessing target data

target

drifter = [1,2,3,4,5,6,7]
t = []
lon = []
lat = []
for i in range(len(target)):
    start_date = datetime(2016, 1, 1)
    y = int(target.iloc[i][0])
    m = int(target.iloc[i][1])
    d = int(target.iloc[i][2])
    target_date = datetime(y, m, d)
    
    h = ((target.iloc[i][3] // 6) + 1) * 0.25
    x = (target_date - start_date).days
    x += h
    x -=1
    t.append(x)
    
    lon_deg = target.iloc[i][5]
    lon_min = target.iloc[i][6]
    
    lat_deg = target.iloc[i][7]
    lat_min = target.iloc[i][8]
    
    _lon = lon_deg + (lon_min / 60)
    _lat = lat_deg + (lat_min / 60)
    
    lon.append(_lon)
    lat.append(_lat)    
    
target["drifter"] = drifter    
target["time"] = t
target["longitude"] = lon
target["latitude"] = lat
target

main = target.iloc[:,9:]
main

water = water.fillna(0)
wind = wind.fillna(0)

main["water_u"] = 0
main["water_v"] = 0
main["wind_u"] = 0
main["wind_v"] = 0
main['d_long'] = 0
main['d_lati'] = 0

# Create a grid of longitude and latitude from the water DataFrame
grid = water[['longitude', 'latitude']].drop_duplicates().values

# Create a KD-Tree for efficient nearest neighbor search
kdtree = cKDTree(grid)

# Function to find the 4 closest vertices for a given point
def find_closest_vertices(row):
    point = np.array([row['longitude'], row['latitude']])
    # Query the KD-Tree to find the 4 closest vertices
    _, indices = kdtree.query(point, k=4)
    return indices

# Apply the function to the main DataFrame to find the 4 closest vertices for each point
main['closest_vertices'] = main.apply(find_closest_vertices, axis=1)

u = []
v = []
for i in range(len(main)):
    x = water.loc[water["time"] == main.iloc[i][1]]
    w_u = []
    w_v = []
    w_d = []
    for j in main.iloc[i][-1]:
        d = ((main.iloc[i][2] - x.iloc[j][1])**2 + (main.iloc[i][3] - x.iloc[j][2])**2) ** 0.5
        w_u.append(x.iloc[j][-2])
        w_v.append(x.iloc[j][-1])
        w_d.append(d)

    # weithged average
    total_weight = sum(w_d)
    weighted_u = sum(w_u[i] * w_d[i] for i in range(len(w_d))) / total_weight
    weighted_v = sum(w_v[i] * w_d[i] for i in range(len(w_d))) / total_weight

    u.append(weighted_u)
    v.append(weighted_v)

main["water_u"] = u
main["water_v"] = v

# Create a grid of longitude and latitude from the water DataFrame
grid2 = wind[['longitude', 'latitude']].drop_duplicates().values

# Create a KD-Tree for efficient nearest neighbor search
kdtree = cKDTree(grid2)

# Function to find the 4 closest vertices for a given point
def find_closest_vertices(row):
    point = np.array([row['longitude'], row['latitude']])
    # Query the KD-Tree to find the 4 closest vertices
    _, indices = kdtree.query(point, k=4)
    return indices

# Apply the function to the main DataFrame to find the 4 closest vertices for each point
main['closest_vertices2'] = main.apply(find_closest_vertices, axis=1)

u = []
v = []
for i in range(len(main)):
    x = wind.loc[wind["time"] == main.iloc[i][1]]
    w_u = []
    w_v = []
    w_d = []
    for j in main.iloc[i][-1]:
        d = ((main.iloc[i][2] - x.iloc[j][1])**2 + (main.iloc[i][3] - x.iloc[j][2])**2) ** 0.5
        w_u.append(x.iloc[j][-2])
        w_v.append(x.iloc[j][-1])
        w_d.append(d)

    # weighted_average
    total_weight = sum(w_d)
    weighted_u = sum(w_u[i] * w_d[i] for i in range(len(w_d))) / total_weight
    weighted_v = sum(w_v[i] * w_d[i] for i in range(len(w_d))) / total_weight

    u.append(weighted_u)
    v.append(weighted_v)

main["wind_u"] = u
main["wind_v"] = v

main = main.iloc[:,:-2]
main

new_main = pd.DataFrame()

# 원본 데이터 프레임의 각 행을 100번 복사하여 추가하고, 조건에 따라 열 값을 설정합니다.
for i in range(len(main)):
    for j in range(100*4):
        row = main.iloc[i].copy()
        row['time'] += j * 0.25
        
        if j > 0:
            row['longitude'] = 0
            row['latitude'] = 0
            row['water_u'] = 0
            row['water_v'] = 0
            row['wind_u'] = 0
            row['wind_v'] = 0
            row['d_long'] = 0
            row['d_lati'] = 0
            
        new_main = pd.concat([new_main, row.to_frame().T], ignore_index=True)

drifter_1 = new_main.loc[new_main["drifter"] == 1].reset_index(drop=True)
drifter_2 = new_main.loc[new_main["drifter"] == 2].reset_index(drop=True)
drifter_3 = new_main.loc[new_main["drifter"] == 3].reset_index(drop=True)
drifter_4 = new_main.loc[new_main["drifter"] == 4].reset_index(drop=True)
drifter_5 = new_main.loc[new_main["drifter"] == 5].reset_index(drop=True)
drifter_6 = new_main.loc[new_main["drifter"] == 6].reset_index(drop=True)
drifter_7 = new_main.loc[new_main["drifter"] == 7].reset_index(drop=True)

drifter_1

result = pd.DataFrame()
drifter_lst = [drifter_1, drifter_2, drifter_3, drifter_4,
              drifter_5, drifter_6, drifter_7]

for df in drifter_lst:

    data = df

    # Create a grid of longitude and latitude from the water DataFrame
    grid = water[['longitude', 'latitude']].drop_duplicates().values

    # Create a KD-Tree for efficient nearest neighbor search
    kdtree = cKDTree(grid)

    # Function to find the 4 closest vertices for a given point
    def find_closest_vertices(row):
        point = np.array([row['longitude'], row['latitude']])
        # Query the KD-Tree to find the 4 closest vertices
        _, indices = kdtree.query(point, k=4)
        return indices

    for i in range(len(data)-1):
        # 예측
        features = [data.iloc[i][4], data.iloc[i][5], data.iloc[i][6], data.iloc[i][7]]
        _d_long = model_d_long.predict([features])
        _d_lati = model_d_lati.predict([features])

        # 예측값 할당
        data.iloc[i][-2] = _d_long
        data.iloc[i][-1] = _d_lati

        # 예측한 값으로 다음 위도 경도 설정
        data.iloc[i+1][2] = data.iloc[i][2] + data.iloc[i][-2]
        data.iloc[i+1][3] = data.iloc[i][3] + data.iloc[i][-1]

        near_lst = find_closest_vertices(data.iloc[i+1])

        x = water.loc[water["time"] == data.iloc[i][1]]
        w_u = []
        w_v = []
        w_d = []
        for j in near_lst:
            d = ((data.iloc[i][2] - x.iloc[j][1])**2 + (data.iloc[i][3] - x.iloc[j][2])**2) ** 0.5
            w_u.append(x.iloc[j][-2])
            w_v.append(x.iloc[j][-1])
            w_d.append(d)

        # weithged average
        total_weight = sum(w_d)
        weighted_u = sum(w_u[i] * w_d[i] for i in range(len(w_d))) / total_weight
        weighted_v = sum(w_v[i] * w_d[i] for i in range(len(w_d))) / total_weight

        data.iloc[i+1][4] = weighted_u
        data.iloc[i+1][5] = weighted_v

    # Create a grid of longitude and latitude from the water DataFrame
    grid2 = wind[['longitude', 'latitude']].drop_duplicates().values

    # Create a KD-Tree for efficient nearest neighbor search
    kdtree = cKDTree(grid2)

    # Function to find the 4 closest vertices for a given point
    def find_closest_vertices(row):
        point = np.array([row['longitude'], row['latitude']])
        # Query the KD-Tree to find the 4 closest vertices
        _, indices = kdtree.query(point, k=4)
        return indices

    for i in range(len(data)-1):
        near_lst = find_closest_vertices(data.iloc[i+1])

        x = wind.loc[wind["time"] == data.iloc[i][1]]
        w_u = []
        w_v = []
        w_d = []
        for j in near_lst:
            d = ((data.iloc[i][2] - x.iloc[j][1])**2 + (data.iloc[i][3] - x.iloc[j][2])**2) ** 0.5
            w_u.append(x.iloc[j][-2])
            w_v.append(x.iloc[j][-1])
            w_d.append(d)

        # weithged average
        total_weight = sum(w_d)
        weighted_u = sum(w_u[i] * w_d[i] for i in range(len(w_d))) / total_weight
        weighted_v = sum(w_v[i] * w_d[i] for i in range(len(w_d))) / total_weight

        data.iloc[i+1][6] = weighted_u
        data.iloc[i+1][7] = weighted_v
    
    result = pd.concat([result, data], ignore_index=True)

result

result.describe()

result.to_csv("data/result_100days.csv", index = False)
